Refactor types.rs into module index with helpers and db modules
- Create types/helpers.rs for non-database helper types:
  - BestChainLocation, NonBestChainLocation
  - TreeRootData, BlockMetadata, BlockWithMetadata

- Transform types.rs into clean module index file:
  - Re-exports db types from db/legacy
  - Re-exports helper types from helpers module
  - Clear documentation of module organization

- Fix non_finalised_state.rs to use height() method instead of field access

- Add Clone derive and doc comment to TreeRootData

Field visibility unchanged - pub(super) fields accessible via re-exports
from parent types module. types.rs is now a proper module index file.
